Add new module for the list feature
